test loader w/ 4 views & pad when read chunk but still less frames than needed & correcter last_failed_video bug & other minor changes

+++ src/datasets/games.py
@@ -77,7 +77,7 @@ class GamesDataset(Dataset):
 
         # Conv 3D expecs float Tensor[C, T, H, W]
         if self.transpose:
-            chunk = chunk.transpose(0, 1).float()
+            chunk = chunk.transpose(0, 1)
 
         return {
                 'video': chunk,
